refactor(mcp): extract prompt handlers to handlers_prompts.rs (lab-kvji.24.1.2)
list_prompts/get_prompt bodies move to inherent list_prompts_impl/get_prompt_impl;
ServerHandler keeps one-line delegators. 'complete' stays in server.rs (Revision 2
budget default). No behavior change.

handlers_prompts.rs: 343 LOC. server.rs 3147 -> 2847. 1423 tests pass.
