+++ crates/lab/src/mcp/handlers_prompts.rs
@@ -1,0 +1,343 @@
+//! Prompt handler bodies (`list_prompts`, `get_prompt`).
+//!
+//! Extracted from `server.rs` (bead `lab-kvji.24.1.2`) as inherent
+//! `impl LabMcpServer` methods. The `ServerHandler` trait impl in
+//! `server.rs` keeps one-line delegators into these `*_impl` methods.
+//!
+//! Named `handlers_prompts` (not `prompts`) because `crate::mcp::prompts`
+//! already owns the builtin prompt definitions this layer calls.
+//!
+//! No behavior change — relocation only.
+
+use std::time::Instant;
+
+use rmcp::ErrorData;
+use rmcp::RoleServer;
+use rmcp::model::{
+    GetPromptRequestParams, GetPromptResult, ListPromptsResult, LoggingLevel,
+    PaginatedRequestParams,
+};
+use rmcp::service::RequestContext;
+use serde_json::Value;
+
+use crate::mcp::context::{auth_context_from_extensions, oauth_upstream_subject_for_request};
+use crate::mcp::logging::DispatchLogOutcome;
+use crate::mcp::server::LabMcpServer;
+
+impl LabMcpServer {
+    pub(crate) async fn list_prompts_impl(
+        &self,
+        _request: Option<PaginatedRequestParams>,
+        context: RequestContext<RoleServer>,
+    ) -> Result<ListPromptsResult, ErrorData> {
+        let start = Instant::now();
+        let subject = self.request_subject_log_tag(&context);
+        tracing::info!(
+            surface = "mcp",
+            service = "labby",
+            action = "list_prompts",
+            subject,
+            "dispatch start"
+        );
+        let mut prompts = crate::mcp::prompts::list_all().prompts;
+
+        if let Some(pool) = self.current_upstream_pool().await {
+            let builtin_names: Vec<String> = prompts
+                .iter()
+                .map(|prompt| prompt.name.to_string())
+                .collect();
+            let builtin_name_refs: Vec<&str> = builtin_names.iter().map(String::as_str).collect();
+            let upstream_prompts = pool.list_upstream_prompts(&builtin_name_refs).await;
+            prompts.extend(upstream_prompts);
+            let auth = auth_context_from_extensions(&context.extensions);
+            if let Some(oauth_subject) =
+                oauth_upstream_subject_for_request(auth, self.request_subject(&context))
+            {
+                let scoped_prompts = pool
+                    .subject_scoped_prompts(
+                        &self.oauth_upstream_configs().await,
+                        oauth_subject.as_ref(),
+                        &builtin_name_refs,
+                    )
+                    .await;
+                prompts.extend(scoped_prompts);
+            }
+        }
+
+        let elapsed_ms = start.elapsed().as_millis();
+        tracing::info!(
+            surface = "mcp",
+            service = "labby",
+            action = "list_prompts",
+            subject,
+            elapsed_ms,
+            "prompt list ok"
+        );
+        self.emit_dispatch_notification(
+            &context,
+            "lab",
+            "list_prompts",
+            elapsed_ms,
+            DispatchLogOutcome::Success,
+        )
+        .await;
+
+        Ok(ListPromptsResult::with_all_items(prompts))
+    }
+
+    pub(crate) async fn get_prompt_impl(
+        &self,
+        request: GetPromptRequestParams,
+        context: RequestContext<RoleServer>,
+    ) -> Result<GetPromptResult, ErrorData> {
+        let start = Instant::now();
+        let subject = self.request_subject_log_tag(&context);
+        tracing::info!(
+            surface = "mcp",
+            service = "labby",
+            action = "get_prompt",
+            subject,
+            prompt = %request.name,
+            "dispatch start"
+        );
+        let args = request
+            .arguments
+            .clone()
+            .unwrap_or_default()
+            .into_iter()
+            .map(|(key, value)| {
+                let string = match value {
+                    Value::String(text) => text,
+                    other => other.to_string(),
+                };
+                (key, string)
+            })
+            .collect();
+
+        if let Some(prompt) = crate::mcp::prompts::get(&self.registry, &request.name, &args) {
+            let elapsed_ms = start.elapsed().as_millis();
+            tracing::info!(
+                surface = "mcp",
+                service = "labby",
+                action = "get_prompt",
+                subject,
+                elapsed_ms,
+                "prompt resolved"
+            );
+            self.emit_dispatch_notification(
+                &context,
+                "lab",
+                "get_prompt",
+                elapsed_ms,
+                DispatchLogOutcome::Success,
+            )
+            .await;
+            return Ok(prompt);
+        }
+
+        if let Some(pool) = self.current_upstream_pool().await
+            && let Some(upstream_name) = pool.find_prompt_owner(&request.name).await
+        {
+            let prompt_name = request.name.clone();
+            tracing::info!(
+                surface = "mcp",
+                service = "labby",
+                action = "get_prompt",
+                prompt = %prompt_name,
+                upstream = %upstream_name,
+                route = "upstream",
+                "dispatch route selected"
+            );
+            let outcome = match pool.get_prompt(&upstream_name, request).await {
+                Some(Ok(result)) => {
+                    let elapsed_ms = start.elapsed().as_millis();
+                    tracing::info!(
+                        surface = "mcp",
+                        service = "labby",
+                        action = "get_prompt",
+                        subject,
+                        prompt = %prompt_name,
+                        upstream = %upstream_name,
+                        elapsed_ms,
+                        "prompt proxy ok"
+                    );
+                    self.emit_dispatch_notification(
+                        &context,
+                        "lab",
+                        "get_prompt",
+                        elapsed_ms,
+                        DispatchLogOutcome::Success,
+                    )
+                    .await;
+                    Ok(result)
+                }
+                Some(Err(message)) => {
+                    let elapsed_ms = start.elapsed().as_millis();
+                    tracing::warn!(
+                        surface = "mcp",
+                        service = "labby",
+                        action = "get_prompt",
+                        prompt = %prompt_name,
+                        upstream = %upstream_name,
+                        elapsed_ms,
+                        kind = "internal_error",
+                        error = %message,
+                        "prompt proxy failed"
+                    );
+                    self.emit_dispatch_notification(
+                        &context,
+                        "lab",
+                        "get_prompt",
+                        elapsed_ms,
+                        DispatchLogOutcome::Failure {
+                            level: LoggingLevel::Error,
+                            kind: "internal_error",
+                        },
+                    )
+                    .await;
+                    Err(ErrorData::internal_error(message, None))
+                }
+                None => {
+                    let elapsed_ms = start.elapsed().as_millis();
+                    tracing::warn!(
+                        surface = "mcp",
+                        service = "labby",
+                        action = "get_prompt",
+                        prompt = %prompt_name,
+                        upstream = %upstream_name,
+                        elapsed_ms,
+                        kind = "not_found",
+                        "upstream not connected for prompt"
+                    );
+                    self.emit_dispatch_notification(
+                        &context,
+                        "lab",
+                        "get_prompt",
+                        elapsed_ms,
+                        DispatchLogOutcome::Failure {
+                            level: LoggingLevel::Warning,
+                            kind: "not_found",
+                        },
+                    )
+                    .await;
+                    Err(ErrorData::invalid_params(
+                        format!("unknown prompt: {prompt_name}"),
+                        None,
+                    ))
+                }
+            };
+            return outcome;
+        }
+
+        let auth = auth_context_from_extensions(&context.extensions);
+        if let Some(oauth_subject) =
+            oauth_upstream_subject_for_request(auth, self.request_subject(&context))
+            && let Some(pool) = self.current_upstream_pool().await
+        {
+            let configs = self.oauth_upstream_configs().await;
+            if let Some(upstream_name) = pool
+                .subject_scoped_prompt_owner(&configs, oauth_subject.as_ref(), &request.name)
+                .await
+                && let Some(config) = configs
+                    .into_iter()
+                    .find(|config| config.name == upstream_name)
+            {
+                let prompt_name = request.name.clone();
+                tracing::info!(
+                    surface = "mcp",
+                    service = "labby",
+                    action = "get_prompt",
+                    prompt = %prompt_name,
+                    upstream = %config.name,
+                    route = "subject_scoped",
+                    oauth_subject = %oauth_subject,
+                    "dispatch route selected"
+                );
+                let outcome = match pool
+                    .subject_scoped_get_prompt(&config, oauth_subject.as_ref(), request)
+                    .await
+                {
+                    Ok(result) => {
+                        let elapsed_ms = start.elapsed().as_millis();
+                        tracing::info!(
+                            surface = "mcp",
+                            service = "labby",
+                            action = "get_prompt",
+                            subject,
+                            oauth_subject = %oauth_subject,
+                            prompt = %prompt_name,
+                            upstream = %config.name,
+                            elapsed_ms,
+                            "subject-scoped prompt proxy ok"
+                        );
+                        self.emit_dispatch_notification(
+                            &context,
+                            "lab",
+                            "get_prompt",
+                            elapsed_ms,
+                            DispatchLogOutcome::Success,
+                        )
+                        .await;
+                        Ok(result)
+                    }
+                    Err(message) => {
+                        let elapsed_ms = start.elapsed().as_millis();
+                        tracing::warn!(
+                            surface = "mcp",
+                            service = "labby",
+                            action = "get_prompt",
+                            prompt = %prompt_name,
+                            upstream = %config.name,
+                            elapsed_ms,
+                            kind = "upstream_error",
+                            error = %message,
+                            "subject-scoped prompt proxy failed"
+                        );
+                        self.emit_dispatch_notification(
+                            &context,
+                            "lab",
+                            "get_prompt",
+                            elapsed_ms,
+                            DispatchLogOutcome::Failure {
+                                level: LoggingLevel::Warning,
+                                kind: "upstream_error",
+                            },
+                        )
+                        .await;
+                        Err(ErrorData::invalid_params(
+                            format!("upstream prompt `{prompt_name}` failed: {message}"),
+                            None,
+                        ))
+                    }
+                };
+                return outcome;
+            }
+        }
+
+        let elapsed_ms = start.elapsed().as_millis();
+        tracing::warn!(
+            surface = "mcp",
+            service = "labby",
+            action = "get_prompt",
+            subject,
+            elapsed_ms,
+            kind = "not_found",
+            "unknown prompt"
+        );
+        self.emit_dispatch_notification(
+            &context,
+            "lab",
+            "get_prompt",
+            elapsed_ms,
+            DispatchLogOutcome::Failure {
+                level: LoggingLevel::Warning,
+                kind: "not_found",
+            },
+        )
+        .await;
+        Err(ErrorData::invalid_params(
+            format!("unknown prompt: {}", request.name),
+            None,
+        ))
+    }
+}
